+++ include/Grapengine/grapengine_pch.hpp
@@ -23,10 +23,11 @@
 #include <variant>
 #include <vector>
 
-//clang-format off
+// clang-format off
 #include "core/ge_config.hpp"
-#include "core/ge_using.hpp"
+#include "core/ge_type_aliases.hpp"
 #include "core/ge_memory.hpp"
-//clang-format on
+#include "core/ge_macros.hpp"
+// clang-format on
 
 #endif
